requirements.txt
anthropic>=0.40.0
google-ads>=25.0.0
supabase>=2.0.0
